+++ apps/app/app/api/queue/portfolio/route.ts
@@ -20,7 +20,10 @@ import { getAddress, isAddress } from "viem";
 import { kafka } from "@/clients/kafka";
 import { prisma } from "@/clients/prisma";
 import { ratelimit } from "@/clients/redis";
-import { MAINNET_CHAINS } from "@/const/chains";
+
+// -----------------------------------------------------------------------------
+// Route
+// -----------------------------------------------------------------------------
 
 export async function POST(request: NextRequest) {
   const id = request.ip ?? "anonymous";
@@ -57,17 +60,9 @@ export async function POST(request: NextRequest) {
     });
   }
 
-  const data = MAINNET_CHAINS.map(chain => {
-    return {
-      topic: "covalent",
-      value: {
-        address: addr,
-        chain_id: chain.id,
-      },
-    };
+  kafka.producer().produce("portfolio", {
+    address: addr,
   });
-
-  kafka.producer().produceMany(data);
 
   return Response.json({
     revalidated: false,
