+++ apps/app/app/api/revalidate/path/route.ts
